use process id and window id for route

+++ content/browser/json_websocket_service_impl.h
@@ -22,7 +22,7 @@ class JsonWebSocketServiceImpl : public mojom::JsonWebSocketService {
  public:
   explicit JsonWebSocketServiceImpl(
       mojo::PendingReceiver<mojom::JsonWebSocketService> receiver,
-      int32_t window_id = -1);
+      int64_t window_id = -1);
   ~JsonWebSocketServiceImpl() override;
 
   // mojom::JsonWebSocketService implementation
@@ -41,13 +41,13 @@ class JsonWebSocketServiceImpl : public mojom::JsonWebSocketService {
 
   // Create with window ID
   static void CreateWithWindowId(
-      int32_t window_id,
+      int64_t window_id,
       mojo::PendingReceiver<mojom::JsonWebSocketService> receiver);
 
  private:
   mojo::Receiver<mojom::JsonWebSocketService> receiver_;
   std::unique_ptr<internal::WebSocketClient> websocket_client_;
-  int32_t window_id_;
+  int64_t window_id_;
 
   base::WeakPtrFactory<JsonWebSocketServiceImpl> weak_factory_{this};
 };
